+++ pkg/apis/application/v1alpha1/chart_types.go
@@ -39,6 +39,7 @@ func NewChartCR() *Chart {
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 // +kubebuilder:storageversion
 // +kubebuilder:subresource:status
+// +kubebuilder:resource:categories=common;giantswarm
 
 type Chart struct {
 	metav1.TypeMeta   `json:",inline"`
@@ -118,7 +119,7 @@ type ChartStatus struct {
 
 type ChartStatusRelease struct {
 	// LastDeployed is the time when the deployed chart was last deployed.
-	LastDeployed metav1.Time `json:"lastDeployed"`
+	LastDeployed metav1.Time `json:"lastDeployed,omitempty"`
 	// Revision is the revision number for this deployed chart.
 	Revision int `json:"revision"`
 	// Status is the status of the deployed chart,