+++ pkg/apis/core/v1alpha1/kvm_cluster_types.go
@@ -5,6 +5,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/giantswarm/apiextensions/pkg/crd"
+	"github.com/giantswarm/apiextensions/pkg/serialization"
 )
 
 const (
@@ -18,6 +19,7 @@ func NewKVMClusterConfigCRD() *v1beta1.CustomResourceDefinition {
 // +genclient
 // +genclient:noStatus
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+// +kubebuilder:resource:categories=giantswarm;kvm
 // +kubebuilder:storageversion
 
 type KVMClusterConfig struct {
@@ -48,12 +50,10 @@ type KVMClusterConfigSpecGuestWorker struct {
 
 // TODO: change MemorySizeGB and StorageSizeGB to resource.Quantity
 type KVMClusterConfigSpecGuestNode struct {
-	ID       string `json:"id"`
-	CPUCores int    `json:"cpuCores,omitempty"`
-	// +kubebuilder:validation:Type=number
-	MemorySizeGB string `json:"memorySizeGB,omitempty"`
-	// +kubebuilder:validation:Type=number
-	StorageSizeGB string `json:"storageSizeGB,omitempty"`
+	ID            string              `json:"id"`
+	CPUCores      int                 `json:"cpuCores,omitempty"`
+	MemorySizeGB  serialization.Float `json:"memorySizeGB,omitempty"`
+	StorageSizeGB serialization.Float `json:"storageSizeGB,omitempty"`
 }
 
 type KVMClusterConfigSpecVersionBundle struct {