+++ pkg/apis/example/v1alpha1/memcached_types.go
@@ -18,6 +18,7 @@ func NewMemcachedConfigCRD() *v1beta1.CustomResourceDefinition {
 // +genclient
 // +genclient:noStatus
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
+// +kubebuilder:resource:categories=example;giantswarm
 // +kubebuilder:storageversion
 
 type MemcachedConfig struct {