+++ pkg/apis/provider/v1alpha1/aws_types.go
@@ -41,6 +41,7 @@ func NewAWSConfigCR() *AWSConfig {
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 // +kubebuilder:storageversion
 // +kubebuilder:subresource:status
+// +kubebuilder:resource:categories=aws;giantswarm
 
 type AWSConfig struct {
 	metav1.TypeMeta   `json:",inline"`