+++ pkg/crd/crd.go
@@ -1,6 +1,7 @@
 package crd
 
 import (
+	"fmt"
 	"io/ioutil"
 	"path/filepath"
 
@@ -16,9 +17,7 @@ import (
 )
 
 const (
-	crdDirectoryV1      = "/config/crd/v1"
-	crdDirectoryV1Beta1 = "/config/crd/v1beta1"
-	crdKind             = "CustomResourceDefinition"
+	crdKind = "CustomResourceDefinition"
 )
 
 var (
@@ -36,23 +35,18 @@ var (
 	}
 )
 
-func find(crdKind schema.GroupVersionKind, crGroup, crKind string) (interface{}, error) {
-	var path string
-	switch crdKind.Version {
-	case v1GroupVersionKind.Version:
-		path = crdDirectoryV1
-	case v1beta1GroupVersionKind.Version:
-		path = crdDirectoryV1Beta1
-	}
+type objectHandler func(data []byte) error
 
+func iterateResources(groupVersionKind schema.GroupVersionKind, handle objectHandler) error {
+	crdDirectory := fmt.Sprintf("/config/crd/%s", groupVersionKind.Version)
 	fs := internal.FS(false)
-	directory, err := fs.Open(path)
+	directory, err := fs.Open(crdDirectory)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return microerror.Mask(err)
 	}
 	files, err := directory.Readdir(0)
 	if err != nil {
-		return nil, microerror.Mask(err)
+		return microerror.Mask(err)
 	}
 
 	for _, info := range files {
@@ -64,72 +58,110 @@ func find(crdKind schema.GroupVersionKind, crGroup, crKind string) (interface{},
 		}
 
 		// Read the file to a string.
-		file, err := fs.Open(filepath.Join(path, info.Name()))
+		file, err := fs.Open(filepath.Join(crdDirectory, info.Name()))
 		if err != nil {
-			return nil, microerror.Mask(err)
+			return microerror.Mask(err)
 		}
 		contents, err := ioutil.ReadAll(file)
 		if err != nil {
-			return nil, microerror.Mask(err)
+			return microerror.Mask(err)
 		}
 
 		// Unmarshal into Unstructured since we don't know if this is a v1 or v1beta1 CRD yet.
 		var object unstructured.Unstructured
 		err = yaml.UnmarshalStrict(contents, &object)
 		if err != nil {
-			return nil, microerror.Mask(err)
+			return microerror.Mask(err)
 		}
-		if object.GetObjectKind().GroupVersionKind() != crdKind {
+		if object.GetObjectKind().GroupVersionKind() != groupVersionKind {
 			continue
 		}
 
-		switch crdKind {
-		case v1beta1GroupVersionKind:
-			var crd v1beta1.CustomResourceDefinition
-			err = yaml.UnmarshalStrict(contents, &crd)
-			if err != nil {
-				return nil, microerror.Mask(err)
-			}
-			if crGroup == crd.Spec.Group && crKind == crd.Spec.Names.Kind {
-				return &crd, nil
-			}
-		case v1GroupVersionKind:
-			var crd v1.CustomResourceDefinition
-			err = yaml.UnmarshalStrict(contents, &crd)
-			if err != nil {
-				return nil, microerror.Mask(err)
-			}
-			if crGroup == crd.Spec.Group && crKind == crd.Spec.Names.Kind {
-				return &crd, nil
-			}
+		if err := handle(contents); err != nil {
+			return microerror.Mask(err)
 		}
 	}
 
-	return nil, microerror.Mask(notFoundError)
+	return nil
+}
+
+var cache []v1.CustomResourceDefinition
+var cacheV1Beta1 []v1beta1.CustomResourceDefinition
+
+// ListV1Beta1 loads all v1beta1 CRDs from the virtual filesystem.
+func ListV1Beta1() ([]v1beta1.CustomResourceDefinition, error) {
+	if cacheV1Beta1 != nil {
+		return cacheV1Beta1, nil
+	}
+
+	handler := func(data []byte) error {
+		var crd v1beta1.CustomResourceDefinition
+		err := yaml.UnmarshalStrict(data, &crd)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+		cacheV1Beta1 = append(cacheV1Beta1, crd)
+		return nil
+	}
+
+	err := iterateResources(v1beta1GroupVersionKind, handler)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	return cacheV1Beta1, nil
+}
+
+// ListV1 loads all v1 CRDs from the virtual filesystem.
+func ListV1() ([]v1.CustomResourceDefinition, error) {
+	if cache != nil {
+		return cache, nil
+	}
+
+	handler := func(data []byte) error {
+		var crd v1.CustomResourceDefinition
+		err := yaml.UnmarshalStrict(data, &crd)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+		cache = append(cache, crd)
+		return nil
+	}
+
+	err := iterateResources(v1GroupVersionKind, handler)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	return cache, nil
 }
 
 // LoadV1Beta1 loads a v1beta1 CRD from the virtual filesystem.
 func LoadV1Beta1(group, kind string) *v1beta1.CustomResourceDefinition {
-	found, err := find(v1beta1GroupVersionKind, group, kind)
+	crds, err := ListV1Beta1()
 	if err != nil {
 		panic(microerror.Mask(err))
 	}
-	crd, ok := found.(*v1beta1.CustomResourceDefinition)
-	if !ok {
-		panic(microerror.Mask(conversionFailedError))
+
+	for _, crd := range crds {
+		if crd.Spec.Names.Kind == kind && crd.Spec.Group == group {
+			return &crd
+		}
 	}
-	return crd
+	panic(microerror.Mask(notFoundError))
 }
 
-// LoadV1Beta1 loads a v1 CRD from the virtual filesystem
-func LoadV1(group, kind string) (out *v1.CustomResourceDefinition) {
-	found, err := find(v1GroupVersionKind, group, kind)
+// LoadV1 loads a v1 CRD from the virtual filesystem
+func LoadV1(group, kind string) *v1.CustomResourceDefinition {
+	crds, err := ListV1()
 	if err != nil {
 		panic(microerror.Mask(err))
 	}
-	crd, ok := found.(*v1.CustomResourceDefinition)
-	if !ok {
-		panic(microerror.Mask(conversionFailedError))
+
+	for _, crd := range crds {
+		if crd.Spec.Names.Kind == kind && crd.Spec.Group == group {
+			return &crd
+		}
 	}
-	return crd
+	panic(microerror.Mask(notFoundError))
 }